pyproject.toml
[project]
name = "madsuiteio"
version = "0.1.0"
description = "A script generating madsuite.github.io"
authors = [
    { name = "Sungho Shin", email = "sushin@mit.edu" },
    { name = "François Pacaud", email = "francois.pacaud@minesparis.psl.eu" },
    { name = "Alexis Montoison", email = "amontoison@anl.gov" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
    "livereload==2.7.1",
    "pybtex==0.24.0",
    "PyYAML==6.0.2",
    "setuptools<=79",
]
